+++ CHANGES.md
@@ -8,6 +8,7 @@
 - win: Don't move the mouse to a relative position if it was not possible to get the current position
 - All: The feature `with_serde` was renamed to `serde`
 - All: Renamed `Key::Layout(char)` to `Key::Unicode(char)` and clearified its docs
+- All: Split off entering raw keycodes into it's own function
 
 ## Added
 - Linux: Support X11 without `xdotools`. Use the experimental feature `x11rb` to test it